Changes as a result of feedback
