+++ google_drive/tests/__init__.py
@@ -1,1 +1,1 @@
-""" Put tests here """
+""" Unit and integration tests for google drive components """
